Refactor: Organize folders and utils for better maintainability and scalability
