+++ src/app/_enums/queue.ts
@@ -22,15 +22,3 @@ export enum QueueType {
   RankedSolo = 'RANKED_SOLO_5x5',
   RankedFlex = 'RANKED_FLEX_SR'
 }
-
-export enum Spell {
-  Summoner1Id = 'summoner1Id',
-  Summoner2Id = 'summoner2Id',
-  Spell1Id = 'spell1Id',
-  Spell2Id = 'spell2Id'
-}
-
-export enum RuneType {
-  MainRune = 'mainRune',
-  SubRune = 'subRune'
-}
