+++ src/app/api/riotGames/riot/newestGameVersion/route.ts
