+++ src/app/api/riotGames/riot/runes/route.ts
@@ -1,5 +1,5 @@
 import { fetchApi } from '@/app/_utils/fetchApi';
-import { riotGamesRoutes } from '@/app/_constants/endpoints';
+import { riotGamesRoutes } from '../../../../_constants/endpoints';
 import { ONE_DAY } from '@/app/_constants/timeUnits';
 import type { TRune } from '@/app/_types/apiTypes/apiTypes';
