+++ src/app/api/riotGames/riot/spectator/route.ts
