+++ src/app/summoners/[region]/[summonerData]/(summaryPage)/_summonerMatchOverview/matchHistory/utils/utils.ts
@@ -1,4 +1,7 @@
-import type { TSummonerDetailedMatchHistory } from '@/app/_types/apiTypes/customApiTypes';
+import type {
+  TSummonerDetailedMatchHistory,
+  TDetailedMatchHistory
+} from '@/app/_types/apiTypes/customApiTypes';
 
 export const getFormattedKillParticipation = (
   summoner: TSummonerDetailedMatchHistory | undefined
@@ -9,8 +12,7 @@ export const getFormattedKillParticipation = (
 export const getFormattedKda = (summoner: TSummonerDetailedMatchHistory | undefined): string => {
   if (summoner?.kills !== 0 && summoner?.deaths === 0 && summoner?.assists !== 0) {
     return 'Perfect';
-  }
-  else {
+  } else {
     return `${summoner?.challenges?.kda.toFixed(2)}:1`;
   }
 }
@@ -21,10 +23,25 @@ export const adjustTeamsOrderBasedOnSummoner = <T>(
 ): Array<T | undefined> | undefined => {
   if (currentSummoner?.teamId === 100) {
     return teams;
-  }
-  else {
+  } else {
     const blueTeam = teams?.[0];
     const redTeam = teams?.[1];
     return [redTeam, blueTeam];
   }
+}
+
+export const getSummonerMinionStats = (
+  summoner: TSummonerDetailedMatchHistory | undefined,
+  match: TDetailedMatchHistory | undefined
+) => {
+  if (summoner && match) {
+    const totalMinions = summoner?.totalMinionsKilled + summoner?.neutralMinionsKilled;
+    const minionsPerMinute = +(totalMinions / (match?.info.gameDuration / 60)).toFixed(1);
+
+    return {
+      minionsPerMinute,
+      totalMinions,
+      minions: summoner?.totalMinionsKilled
+    };
+  };
 }
