+++ src/app/summoners/[region]/[summonerData]/(summaryPage)/_summonerRank/SummonerRank.tsx
@@ -6,11 +6,11 @@ import { useAppSelector } from '@/app/_hooks/useReduxHooks';
 import { useQuery } from '@tanstack/react-query';
 import { fetchApi } from '@/app/_utils/fetchApi';
 import { riotGamesRoutes } from '@/app/_constants/endpoints';
-import { findQueueTypeData, calculatePercentage } from '@/app/_utils/matchStats';
-import { formatTierName, getRankedEmblem } from '@/app/_utils/rank';
-import SummonerRankSkeleton from './SummonerRankSkeleton';
+import { calculatePercentage } from '@/app/_utils/utils';
+import { formatTierName, getRankedEmblem, findQueueTypeData } from '@/app/_utils/rank';
 import type { TSummonerRank } from '@/app/_types/apiTypes/apiTypes';
-import { QueueType } from '@/app/_enums/match';
+import { QueueType } from '@/app/_enums/queue';
+import SummonerRankSkeleton from './SummonerRankSkeleton';
 
 type Props = {
   queueType: QueueType;
@@ -28,7 +28,7 @@ const SummonerRank = ({ queueType, smallDataStyle }: Props) => {
     refetchOnWindowFocus: false
   });
 
-  const rankedData: TSummonerRank | undefined = findQueueTypeData(fetchedSummonerRanksData, queueType);
+  const rankedData = findQueueTypeData(fetchedSummonerRanksData, queueType);
 
   const totalPlayedGames = rankedData ? rankedData?.wins + rankedData?.losses : 0;
   const winRate = calculatePercentage(rankedData?.wins, totalPlayedGames);
